인기사용자 리스트 완료 댓글 작성 완료 댓글 목록 불러오기 작업 시작

+++ src/main/java/com/dto/UserInfo.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UserInfo {
-	private long userSeq;
+	private int userSeq;
 	private String userId;
 	private String userNick;
 	private int roles;
